+++ src/main/kotlin/lotto/Application.kt
@@ -3,6 +3,7 @@ package lotto
 import lotto.input.PurchaseAmountInputReader
 import lotto.calculation.TicketCalculator
 import lotto.calculation.TicketGenerator
+import lotto.display.TicketDisplay
 
 fun main() {
 
@@ -14,5 +15,8 @@ fun main() {
 
     // Generate random tickets as Lotto objects
     val tickets = TicketGenerator().generateTickets(numberOfTickets)
+
+    // Display the generated Lotto tickets
+    TicketDisplay().displayTickets(numberOfTickets, tickets)
 
 }